Move and combine transformer tests
